% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-rmd.R
\name{download_rmd_button}
\alias{download_rmd_button}
\title{Generate an HTML widget to download input Rmd file}
\usage{
download_rmd_button(
  input = NULL,
  text = "Download Rmd",
  ...,
  align = "right",
  class = "button",
  aside = TRUE,
  embed = xfun::embed_file
)
}
\arguments{
\item{input}{Filename of the input. If \code{NULL}, the function automatically
detects the name of the knitting Rmd file.}

\item{text}{The text for the hyperlink.}

\item{...}{Arguments passed to the \code{embed} function.}

\item{align}{Align the button by CSS's \code{text-align} attribute.
This feature is disabled when aside is \code{FALSE}}

\item{embed}{A function to embed file(s).
One of \code{xfun::embed_file}, \code{xfun::embed_files}, or \code{xfun::embed_dir}.}
}
\value{
\code{shiny.tag} class object.
}
\description{
The button can self-contain the input data by base64 encoding.
}
\examples{
input <- tempfile()
download_rmd_button(input)

input <- tempdir()
download_rmd_button(input, embed = xfun::embed_dir)

}
