% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameworks.R
\docType{data}
\name{frameworks}
\alias{frameworks}
\title{Lightweight CSS frameworks meta data}
\format{A list}
\usage{
frameworks
}
\description{
To see the names of the frameworks, run \code{names(frameworks)}.
To see the names of themes of a framework, say \code{"mini"}, run
\code{names(frameworks$mini$stylesheet)}
}
\keyword{datasets}
