% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hook-results-folding.R
\name{set_results_folding}
\alias{set_results_folding}
\title{Folding chunk results in R Markdown documents}
\usage{
set_results_folding(
  results_folding = c("none", "show", "hide"),
  knit_hook_source = knitr::knit_hooks$get("source")
)
}
\arguments{
\item{results_folding}{Setup results folding by a string, \code{"none"}, \code{"show"},
or \code{"hide"}. If \code{"none"}, the feature appears to be disabled, however,
can be opt-in per chunk by specifying \code{"show"} or \code{"hide"} on the
\code{results.folding} chunk option.}

\item{knit_hook_source}{A base hook for the source. By default, the function
seeks for the current hook and overlay it.}
}
\value{
Invisible \code{NULL}
}
\description{
The folding covers entire results including figures and tables. Note that
this feature only works with HTML outputs.
}
\examples{
set_results_folding()

# New hooks
knitr::knit_hooks$get()[c("source", "results.folding")]

}
