% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{spec_includes}
\alias{spec_includes}
\title{cdn, framework, and theme are currently ignored}
\usage{
spec_includes(includes = list(), katex = TRUE)
}
\description{
cdn, framework, and theme are currently ignored
}
