% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini_document.R
\name{mini_document}
\alias{mini_document}
\title{Convert to an HTML document powered by the 'mini.css' framework.}
\usage{
mini_document(
  code_folding = c("none", "show", "hide"),
  pandoc_args = NULL,
  keep_md = FALSE,
  includes = list(),
  ...
)
}
\arguments{
\item{code_folding}{Enable document readers to toggle the display of R code
chunks. Specify \code{"none"} to display all code chunks (assuming
they were knit with \code{echo = TRUE}). Specify \code{"hide"} to hide all R
code chunks by default (users can show hidden code chunks either
individually or document-wide). Specify \code{"show"} to show all R code
chunks by default.}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{keep_md}{Keep the markdown file generated by knitting.}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link[rmarkdown]{includes}} function).}

\item{...}{Arguments passed to \code{rmarkdown::html_document}}
}
\description{
Convert to an HTML document powered by the 'mini.css' framework.
}
